+++ R/data.R
@@ -4,35 +4,35 @@
 #' Voxel coordinates for an organic shape
 #' 
 #' @examples
-#' head(coords_organic)
-#' cubes <- isocubesGrob(coords_organic, size = 2)
+#' head(obj_organic)
+#' cubes <- isocubesGrob(obj_organic, size = 2)
 #' grid::grid.newpage()
 #' grid::grid.draw(cubes)
 #' @family datasets
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-"coords_organic"
+"obj_organic"
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' Voxel coordinates for a sphere
 #' 
 #' @examples
-#' head(coords_sphere)
-#' cubes <- isocubesGrob(coords_sphere, size = 2)
+#' head(obj_sphere)
+#' cubes <- isocubesGrob(obj_sphere, size = 2)
 #' grid::grid.newpage()
 #' grid::grid.draw(cubes)
 #' @family datasets
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-"coords_sphere"
+"obj_sphere"
 
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #' Voxel coordinates for a sphere
 #' 
 #' @examples
-#' head(coords_letter)
-#' cubes <- isocubesGrob(coords_letter, size = 5, y = 0.05)
+#' head(obj_letter)
+#' cubes <- isocubesGrob(obj_letter, size = 5, y = 0.05)
 #' grid::grid.newpage()
 #' grid::grid.draw(cubes)
 #' @family datasets
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-"coords_letter"
+"obj_letter"